Add display tab and subtabs
