+++ src/admin-views/settings-tabs/general/general-debugging.php
@@ -9,7 +9,7 @@
 $tec_events_general_debugging = [
 	'tec-events-settings-general-debugging-title' => [
 		'type' => 'html',
-		'html' => '<h3 id="tec-settings-general-debugging">' . esc_html_x( 'Debugging', 'Title for the debugging section of the general settings.', 'the-events-calendar' ) . '</h3>',
+		'html' => '<h3 id="tec-settings-general-debugging" class="tec_settings__section-header">' . esc_html_x( 'Debugging', 'Title for the debugging section of the general settings.', 'the-events-calendar' ) . '</h3>',
 	],
 	'tec-troubleshooting-infobox-start'           => [
 		'type' => 'html',
@@ -55,12 +55,17 @@ $tec_events_general_debugging = [
 	],
 ];
 
-$debugging = new Tribe__Settings_Tab(
-	'debugging',
+$general_debugging = new Tribe__Settings_Tab(
+	'general-debugging-tab',
 	esc_html__( 'Debugging', 'the-events-calendar' ),
 	[
-		'priority' => 15,
-		'fields'   => apply_filters( 'tribe_general_settings_debugging_section', $tec_events_general_debugging ),
+		'priority' => 0.15,
+		'fields'   => apply_filters(
+			'tribe_general_settings_debugging_section',
+			$tec_events_general_debugging
+		),
 		'parent'   => 'general',
 	]
 );
+
+do_action( 'tec_events_display_general_debugging_tab', $general_debugging );
