+++ src/admin-views/settings-tabs/general/general-viewing.php
@@ -12,7 +12,7 @@ $tec = Tribe__Events__Main::instance();
 $tec_events_general_viewing_fields = [
 	'tec-events-settings-general-viewing-title' => [
 		'type' => 'html',
-		'html' => '<h3 id="tec-settings-general-viewing">' . esc_html_x( 'Viewing', 'Title for the viewing section of the general settings.', 'the-events-calendar' ) . '</h3>',
+		'html' => '<h3 id="tec-settings-general-viewing" class="tec_settings__section-header">' . esc_html_x( 'Viewing', 'Title for the viewing section of the general settings.', 'the-events-calendar' ) . '</h3>',
 	],
 	'unpretty-permalinks-url'                   => [
 		'type'        => 'wrapped_html',
@@ -89,11 +89,11 @@ $tec_events_general_viewing_fields = [
 	],
 ];
 
-$tec_events_general_viewing = new Tribe__Settings_Tab(
-	'viewing',
+new Tribe__Settings_Tab(
+	'general-viewing-tab',
 	esc_html__( 'Viewing', 'the-events-calendar' ),
 	[
-		'priority' => 1,
+		'priority' => 0.0,
 		'fields'   => apply_filters( 'tribe_general_settings_viewing_section', $tec_events_general_viewing_fields ),
 		'parent'   => 'general',
 	]
