+++ src/views/v2/latest-past/event/venue.php
@@ -20,7 +20,7 @@ if ( ! $event->venues->count() ) {
 	return;
 }
 
-$separator            = esc_html_x( ', ', 'Address separator', 'events-calendar-pro' );
+$separator            = esc_html_x( ', ', 'Address separator', 'tribe-events-calendar-pro' );
 $venue                = $event->venues[0];
 $append_after_address = array_filter( array_map( 'trim', [ $venue->city, $venue->state_province, $venue->state, $venue->province ] ) );
 $address              = $venue->address . ( $venue->address && $append_after_address ? $separator : '' );
